MLFlix-Recommender/Demo2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
jupyter==1.0.0
notebook==7.0.6
pytest==7.4.3
httpx==0.25.2
email-validator==2.1.0.post1
bcrypt==4.0.1
cryptography==41.0.5